+++ das_client/lib/app/pages/journey/train_journey/widgets/header/time_container.dart
@@ -1,6 +1,9 @@
+import 'package:das_client/app/bloc/train_journey_cubit.dart';
 import 'package:das_client/app/widgets/das_text_styles.dart';
 import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
 import 'package:flutter/material.dart';
+import 'package:das_client/model/journey/journey.dart';
+import 'package:intl/intl.dart';
 
 class TimeContainer extends StatelessWidget {
   const TimeContainer({super.key});
@@ -21,21 +24,67 @@ class TimeContainer extends StatelessWidget {
         height: 112.0,
         child: Column(
           mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-          crossAxisAlignment: CrossAxisAlignment.center,
+          crossAxisAlignment: CrossAxisAlignment.end,
           children: [
-            Text('05:43:00', style: DASTextStyles.xLargeBold),
+            Flexible(child: _currentTime()),
             _divider(),
-            Text('+00:01:30', style: DASTextStyles.xLargeLight),
+            Flexible(child: _punctualityDisplay(context)),
           ],
         ),
       ),
     );
   }
+}
 
-  Widget _divider() {
-    return const Padding(
-      padding: EdgeInsets.symmetric(vertical: sbbDefaultSpacing * 0.5),
-      child: Divider(height: 1.0, color: SBBColors.cloud),
-    );
-  }
+Widget _divider() {
+  return const Padding(
+    padding: EdgeInsets.symmetric(vertical: sbbDefaultSpacing * 0.5),
+    child: Divider(height: 1.0, color: SBBColors.cloud),
+  );
+}
+
+Widget _punctualityDisplay(BuildContext context) {
+  final bloc = context.trainJourneyCubit;
+
+  return StreamBuilder<Journey?>(
+    stream: bloc.journeyStream,
+    builder: (context, snapshot) {
+      if (!snapshot.hasData || snapshot.data == null || snapshot.data!.metadata.delay == null) {
+        return Padding(
+          padding:
+              const EdgeInsets.fromLTRB(sbbDefaultSpacing * 0.5, 0.0, sbbDefaultSpacing * 0.5, sbbDefaultSpacing * 0.5),
+          child: Text('+00:00', style: DASTextStyles.xLargeLight),
+        );
+      }
+
+      final Journey journey = snapshot.data!;
+      final Duration delay = journey.metadata.delay!;
+
+      final String minutes = NumberFormat('00').format(delay.inMinutes.abs() % 60);
+      final String seconds = NumberFormat('00').format(delay.inSeconds.abs() % 60);
+      final String formattedDuration = '${delay.isNegative ? '-' : '+'}$minutes:$seconds';
+
+      return Padding(
+        padding:
+            const EdgeInsets.fromLTRB(sbbDefaultSpacing * 0.5, 0.0, sbbDefaultSpacing * 0.5, sbbDefaultSpacing * 0.5),
+        child: Text(formattedDuration, style: DASTextStyles.xLargeLight),
+      );
+    },
+  );
+}
+
+StreamBuilder _currentTime() {
+  return StreamBuilder(
+    stream: Stream.periodic(const Duration(milliseconds: 200)),
+    builder: (context, snapshot) {
+      return Padding(
+        padding:
+            const EdgeInsets.fromLTRB(sbbDefaultSpacing * 0.5, sbbDefaultSpacing * 0.5, sbbDefaultSpacing * 0.5, 0),
+        child: Text(
+          DateFormat('HH:mm:ss').format(DateTime.now().toLocal()),
+          style: DASTextStyles.xLargeBold,
+        ),
+      );
+    },
+  );
 }