+++ das_client/lib/model/journey/additional_speed_restriction.dart
@@ -1,7 +1,7 @@
 import 'package:das_client/model/journey/base_data.dart';
 
 class AdditionalSpeedRestriction {
-  AdditionalSpeedRestriction(
+  const AdditionalSpeedRestriction(
       {required this.kmFrom, required this.kmTo, required this.orderFrom, required this.orderTo, this.speed});
 
   final double kmFrom;