+++ das_client/lib/model/journey/cab_signaling.dart
@@ -2,7 +2,7 @@ import 'package:das_client/model/journey/base_data.dart';
 import 'package:das_client/model/journey/datatype.dart';
 
 class CABSignaling extends BaseData {
-  CABSignaling({
+  const CABSignaling({
     required super.order,
     required super.kilometre,
     this.isStart = false,