+++ das_client/lib/model/journey/metadata.dart
@@ -1,18 +1,21 @@
 import 'package:das_client/model/journey/additional_speed_restriction.dart';
 import 'package:das_client/model/journey/base_data.dart';
+import 'package:das_client/model/journey/bracket_station_segment.dart';
 import 'package:das_client/model/journey/break_series.dart';
 import 'package:das_client/model/journey/service_point.dart';
-import 'package:das_client/model/journey/track_equipment.dart';
+import 'package:das_client/model/journey/track_equipment_segment.dart';
 
 class Metadata {
-  Metadata({
+  const Metadata({
     this.nextStop,
     this.currentPosition,
     this.routeStart,
     this.routeEnd,
+    this.delay,
     this.breakSeries,
     this.additionalSpeedRestrictions = const [],
     this.nonStandardTrackEquipmentSegments = const [],
+    this.bracketStationSegments = const [],
     this.availableBreakSeries = const {},
   });
 
@@ -21,7 +24,9 @@ class Metadata {
   final List<AdditionalSpeedRestriction> additionalSpeedRestrictions;
   final BaseData? routeStart;
   final BaseData? routeEnd;
+  final Duration? delay;
   final List<NonStandardTrackEquipmentSegment> nonStandardTrackEquipmentSegments;
+  final List<BracketStationSegment> bracketStationSegments;
   final BreakSeries? breakSeries;
   final Set<BreakSeries> availableBreakSeries;
 }