+++ das_client/lib/model/journey/tram_area.dart
@@ -2,13 +2,13 @@ import 'package:das_client/model/journey/base_data.dart';
 import 'package:das_client/model/journey/datatype.dart';
 
 class TramArea extends BaseData {
-  TramArea(
-      {required super.order,
-      required super.kilometre,
-      required this.endKilometre,
-      required this.amountTramSignals,
-      super.speedData})
-      : super(type: Datatype.tramArea);
+  const TramArea({
+    required super.order,
+    required super.kilometre,
+    required this.endKilometre,
+    required this.amountTramSignals,
+    super.speedData,
+  }) : super(type: Datatype.tramArea);
 
   final double endKilometre;
   final int amountTramSignals;