+++ das_client/test/model/journey/track_equipment_test.dart
@@ -1,4 +1,4 @@
-import 'package:das_client/model/journey/track_equipment.dart';
+import 'package:das_client/model/journey/track_equipment_segment.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {